Refactored AppView to not be a HOC so that hooks are available for use in App
