+++ src/site/pages/digital/src/containers/App/index.tsx
@@ -1,29 +1,8 @@
-import {createRef} from "react";
-
-import {InteractionTool}    from "core/tools/InteractionTool";
-import {PanTool}            from "core/tools/PanTool";
-import {RotateTool}         from "core/tools/RotateTool";
-import {TranslateTool}      from "core/tools/TranslateTool";
-import {WiringTool}         from "core/tools/WiringTool";
-import {SplitWireTool}      from "core/tools/SplitWireTool";
-import {SelectionBoxTool}   from "core/tools/SelectionBoxTool";
-
-import {SelectAllHandler}     from "core/tools/handlers/SelectAllHandler";
-import {FitToScreenHandler}   from "core/tools/handlers/FitToScreenHandler";
-import {DuplicateHandler}     from "core/tools/handlers/DuplicateHandler";
-import {DeleteHandler}        from "core/tools/handlers/DeleteHandler";
-import {SnipWirePortsHandler} from "core/tools/handlers/SnipWirePortsHandler";
-import {DeselectAllHandler}   from "core/tools/handlers/DeselectAllHandler";
-import {SelectionHandler}     from "core/tools/handlers/SelectionHandler";
-import {SelectPathHandler}    from "core/tools/handlers/SelectPathHandler";
-import {UndoHandler}          from "core/tools/handlers/UndoHandler";
-import {RedoHandler}          from "core/tools/handlers/RedoHandler";
-import {CopyHandler}          from "core/tools/handlers/CopyHandler";
-import {PasteHandler}         from "core/tools/handlers/PasteHandler";
-
 import {CircuitMetadataBuilder} from "core/models/CircuitMetadata";
 
-import {SetCircuitSaved} from "shared/state/CircuitInfo";
+import {DigitalCircuitInfo} from "digital/utils/DigitalCircuitInfo";
+
+import {CircuitInfoHelpers} from "shared/utils/CircuitInfoHelpers";
 
 import {ContextMenu}     from "shared/containers/ContextMenu";
 import {Header}          from "shared/containers/Header";
@@ -34,9 +13,6 @@ import {SelectionPopup}       from "shared/containers/SelectionPopup";
 import {PositionModule}       from "shared/containers/SelectionPopup/modules/PositionModule";
 
 import {DigitalPaste} from "site/digital/utils/DigitalPaste";
-import {Setup}        from "site/digital/utils/CircuitInfo/Setup";
-
-import {AppStore} from "site/digital/state";
 
 import {DigitalItemNav}         from "site/digital/containers/DigitalItemNav";
 import {ICDesigner}             from "site/digital/containers/ICDesigner";
@@ -73,66 +49,48 @@ const exampleCircuits = exampleConfig.examples.map((example) =>
         .build()
 );
 
-export const App = ((store: AppStore) => {
-    const canvas = createRef<HTMLCanvasElement>();
+type Props = {
+    info: DigitalCircuitInfo;
+    helpers: CircuitInfoHelpers;
+    canvas: React.RefObject<HTMLCanvasElement>;
+}
+export const App = ({info, helpers, canvas}: Props) => {
+    return (
+        <div className="App">
+            <SideNav helpers={helpers}
+                        exampleCircuits={exampleCircuits} />
 
-    // Setup circuit and get the CircuitInfo and helpers
-    const [info, helpers] = Setup(
-        store, canvas,
-        new InteractionTool([
-            SelectAllHandler, FitToScreenHandler, DuplicateHandler,
-            DeleteHandler, SnipWirePortsHandler, DeselectAllHandler,
-            SelectionHandler, SelectPathHandler, RedoHandler, UndoHandler,
-            CopyHandler, PasteHandler((data) => DigitalPaste(data, info))
-        ]),
-        PanTool, RotateTool,
-        TranslateTool, WiringTool,
-        SplitWireTool, SelectionBoxTool
-    );
+            <div className="App__container">
+                <Header img="img/icons/logo.svg"
+                        helpers={helpers} />
 
-    info.history.addCallback(() => {
-        store.dispatch(SetCircuitSaved(false));
-    });
+                <main>
+                    <MainDesigner info={info} canvas={canvas} />
 
+                    <DigitalItemNav info={info} />
 
-    return function AppView() {
-        return (
-            <div className="App">
-                <SideNav helpers={helpers}
-                         exampleCircuits={exampleCircuits} />
+                    <SelectionPopup info={info}
+                                    modules={[PositionModule, InputCountModule,
+                                              SelectPortCountModule,
+                                              DecoderInputCountModule,
+                                              OutputCountModule, SegmentCountModule,
+                                              ClockFrequencyModule,
+                                              ColorModule, TextColorModule,
+                                              BusButtonModule, CreateICButtonModule,
+                                              ViewICButtonModule, ClockSyncButtonModule]} />
 
-                <div className="App__container">
-                    <Header img="img/icons/logo.svg"
-                            helpers={helpers} />
-
-                    <main>
-                        <MainDesigner info={info} canvas={canvas} />
-
-                        <DigitalItemNav info={info} />
-
-                        <SelectionPopup info={info}
-                                        modules={[PositionModule, InputCountModule,
-                                                  SelectPortCountModule,
-                                                  DecoderInputCountModule,
-                                                  OutputCountModule, SegmentCountModule,
-                                                  ClockFrequencyModule,
-                                                  ColorModule, TextColorModule,
-                                                  BusButtonModule, CreateICButtonModule,
-                                                  ViewICButtonModule, ClockSyncButtonModule]} />
-
-                        <ContextMenu info={info}
-                                     paste={(data) => DigitalPaste(data, info)} />
-                    </main>
-                </div>
-
-                <ICDesigner mainInfo={info} />
-                <ICViewer mainInfo={info} />
-
-                <QuickStartPopup />
-                <KeyboardShortcutsPopup />
-
-                <LoginPopup />
+                    <ContextMenu info={info}
+                                    paste={(data) => DigitalPaste(data, info)} />
+                </main>
             </div>
-        );
-    }
-});
+
+            <ICDesigner mainInfo={info} />
+            <ICViewer mainInfo={info} />
+
+            <QuickStartPopup />
+            <KeyboardShortcutsPopup />
+
+            <LoginPopup />
+        </div>
+    );
+}
